requirements.txt
streamlit
pandas
numpy
pymongo
python-dotenv
sentence-transformers
google-generativeai